+++ embeduptime/info.json
@@ -3,11 +3,11 @@
         "max"
     ],
     "name": "EmbedUptime",
-    "install_msg": "Thanks for installing. Once you load this cog, it will replace the uptime in core with this one, on unload you will have the core one back.\nYou can toggle replies off by using `[p]uptimeset`.\nFeel free to check out the docs for this cog <https://maxbooiii.gitbook.io/maxcogs/cogs/embeduptime>",
+    "install_msg": "Thanks for installing. Once you load this cog, it will replace the uptime in core with this one, on unload you will have the core one back. **Important note**, this cog support slash command, once you sync the tree it will show up under `/` on discord.\nFeel free to check out the docs for this cog <https://maxbooiii.gitbook.io/maxcogs/cogs/embeduptime>",
     "description": "Shows current uptime of the bot.",
     "short": "Shows current uptime of the bot.",
     "hidden": false,
-    "min_bot_version": "3.5.0.dev1",
+    "min_bot_version": "3.5.0",
     "tags": [
         "uptime"
     ],
